inference/test/requirements.txt
boto3==1.34.63
botocore==1.34.63
Pillow==10.3.0
matplotlib==3.7.5
opencv-python==4.9.0.80
python-dotenv==1.0.1
numpy>=1.21.0,<2.0.0